refactor(contenido-categoria) creacion-del-paquete-superclass
