+++ .gitignore
@@ -4,7 +4,6 @@ build/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
-.env.log
 
 ### STS ###
 .apt_generated
@@ -36,3 +35,8 @@ out/
 
 ### VS Code ###
 .vscode/
+
+# Environment variables
+.env
+.env.local
+.env.*.local
