+++ src/main/java/com/cesde/Chatbot/model/Sugerencia.java
@@ -1,5 +1,7 @@
 package com.cesde.Chatbot.model;
 
+import com.cesde.Chatbot.superclass.Contenido;
+
 import jakarta.persistence.Entity;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
