+++ src/main/resources/application.properties
@@ -1,14 +1,14 @@
 spring.application.name=Chatbot
 
 #Puerto de la app
-server.port = 443
+server.port = ${SERVER_PORT:443}
 
 # ========================================
 # MYSQL DATABASE CONFIGURATION
 # ========================================
-spring.datasource.url=jdbc:mysql://localhost:3306/chatbot?useSSL=false&serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=Pancito11ProMax...
+spring.datasource.url=jdbc:mysql://localhost:3306/${DB_NAME:mydb}?useSSL=false&serverTimezone=UTC
+spring.datasource.username=${DB_USER:root}
+spring.datasource.password=${DB_PASSWORD:password}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 # ========================================
@@ -40,3 +40,11 @@ logging.level.org.springdoc=DEBUG
 logging.level.io.swagger=DEBUG
 logging.level.org.hibernate.SQL=DEBUG
 logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
+
+# ========================================
+# ACTUATOR CONFIGURATION FOR HEALTH CHECK
+# ========================================
+management.endpoints.web.exposure.include=health,info,metrics
+management.endpoint.health.show-details=always
+management.health.db.enabled=true
+management.metrics.export.prometheus.enabled=false
